+++ src/selector/focusable.supports.js
@@ -5,7 +5,6 @@ import canFocusVideoWithoutControls from '../supports/focus-video-without-contro
 import canFocusTable from '../supports/focus-table';
 import canFocusFieldset from '../supports/focus-fieldset';
 import canFocusSummary from '../supports/focus-summary';
-import cssShadowPiercingDeepCombinator from '../supports/css-shadow-piercing-deep-combinator';
 
 export default memorizeResult(function() {
   return {
@@ -14,6 +13,5 @@ export default memorizeResult(function() {
     canFocusTable: canFocusTable(),
     canFocusFieldset: canFocusFieldset(),
     canFocusSummary: canFocusSummary(),
-    cssShadowPiercingDeepCombinator: cssShadowPiercingDeepCombinator(),
   };
 });